+++ apps/web/src/components/auth/ForgotPasswordForm.tsx
@@ -9,11 +9,11 @@
 
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { Button, Input } from "@shared";
+import { Button, Input } from "@/components/ui/forms";
 import { ServerErrorBanner } from "@/components/shared";
 import { useForgotPasswordMutation } from "@shared/api/authApi";
-import { useAuthForm } from "@shared";
-import { validateForgotPasswordForm } from "@shared";
+import { useAuthForm } from "@/components/ui/forms";
+import { validateForgotPasswordForm } from "@/components/ui/forms";
 
 interface ForgotPasswordFormData {
     email: string;
