+++ apps/web/src/components/auth/RegisterForm.tsx
@@ -12,15 +12,15 @@
 import React from "react";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { Button, Input, FormSelect } from "@shared";
+import { Button, Input, FormSelect } from "@/components/ui/forms";
 import { ServerErrorBanner } from "@/components/shared";
 import { useRegisterMutation } from "@shared/api/authApi";
 import { loginSuccess, loginFailure } from "@shared/store/authSlice";
-import { useAuthForm, USER_ROLES } from "@shared";
-import { validateRegisterForm } from "@shared";
+import { useAuthForm, USER_ROLES } from "@/components/ui/forms";
+import { validateRegisterForm } from "@/components/ui/forms";
 import type { AppDispatch } from "@shared/store";
 import type { RegisterCredentials, UserRole } from "@shared/types/auth";
-import type { SelectOption } from "@shared";
+import type { SelectOption } from "@/components/ui/forms";
 
 interface RegisterFormData {
     email: string;
