--- apps/web/src/components/forms/index.ts
@@ -1,1 +1,0 @@
-// Components migrados a @shared
