--- packages/shared/src/components/index.ts
@@ -1,10 +1,0 @@
-/**
- * Exports centralizados de todos los componentes reutilizables
- * Facilita imports desde other packages y apps
- * Mantiene arquitectura limpia y escalable
- * 
- * @author Frontend Team
- * @since v1.0.0
- */
-
-export * from './forms';
